requirements.txt
Flask==2.0.1
requests==2.26.0
numpy==1.21.1
opencv-python==4.5.3.56
line-bot-sdk==1.18.0
SQLAlchemy==1.3.22
gunicorn==20.0.4